# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Basic API",
        "description": "L5 Swagger API with Laravel basic",
        "contact": {
            "email": "nguyen.an.thoi.212@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/laravel_basic/an.thoi.laravel/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/friend/{friend_id}/request-add-friend": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Request Add Friend",
                "description": "Request Add Friend",
                "operationId": "requestAddFriend",
                "parameters": [
                    {
                        "name": "friend_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request Add Friend successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/list-add-friend/{friend_request_id}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Accept Add Friend",
                "description": "Accept Add Friend",
                "operationId": "acceptAddFriend",
                "parameters": [
                    {
                        "name": "friend_request_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accept Add Friend successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/list-add-friend": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get List Request Add Friend",
                "description": "Get List Request Add Friend",
                "operationId": "GetListAddFriend",
                "responses": {
                    "200": {
                        "description": "Get List Add Friend Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/list-friend": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get List Friend",
                "description": "Get List Friend",
                "operationId": "GetListFriend",
                "responses": {
                    "200": {
                        "description": "Get List Friend Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/{friend_list_id}/friend-profile": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Get Friend Profile",
                "description": "Get Friend Profile",
                "operationId": "GetFriendProfile",
                "parameters": [
                    {
                        "name": "friend_list_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Friend Profile Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "A short description for security scheme",
            "name": "Authorization",
            "in": "header"
        }
    }
}